+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 - Fixed DeletedAt field in subscriber model
 - Fixed Dockerfile
+- Updated platform-events dependency to v0.0.11
 
 ## [0.0.5] - 2023-07-06